+++ digi/xbee/recovery.py
@@ -1,4 +1,4 @@
-# Copyright 2019, 2020, Digi International Inc.
+# Copyright 2019-2021, Digi International Inc.
 #
 # This Source Code Form is subject to the terms of the Mozilla Public
 # License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -30,7 +30,8 @@ from digi.xbee.util import utils
 
 SUPPORTED_HARDWARE_VERSIONS = (HardwareVersion.XBEE3.code,
                                HardwareVersion.XBEE3_SMT.code,
-                               HardwareVersion.XBEE3_TH.code)
+                               HardwareVersion.XBEE3_TH.code,
+                               HardwareVersion.XBEE3_RR.code)
 
 _BAUDRATE_KEY = "baudrate"
 _PARITY_KEY = "parity"